\name{archeofrag-package}
\alias{archeofrag-package}
\alias{archeofrag}
\docType{package}
\title{
  Archeofrag: tools to analyse the fragmentation of archeological objects 
}
\description{
    The Archeofrag package implements a comprehensive procedure to 
    analyse fragmentation processes in archeology, based on the 
    fragments of archeological objects scattered in stratigraphic 
    layers and connected by physical "refittings" relationships. 
    Graphs are used to model these situations. The package is mainly 
    based on the igraph package for graph analysis. It offers functions 
    to manipulate fragment datasets, to measure the properties of the 
    graphs, and to generate artificial cases to compare with 
    archaeological observations. An empirical dataset is also provided 
    as an example.
}
\details{
    \tabular{ll}{
        Package: \tab archeofrag\cr
        Type: \tab Package\cr
        Version: \tab 0.5.4\cr
        Date: \tab 2020-09-30\cr
        License: \tab GPL-2\cr
    }
}
\author{Sebastien Plutniak
Maintainer: Sebastien Plutniak <sebastien.plutniak at posteo.net>
}
\keyword{ package }
\seealso{
\code{\link[igraph:igraph-package]{igraph}} 
\code{\link[RBGL:RBGL.overview]{RBGL}} 
}
\examples{}
