\name{frag.layers.cohesion}
\alias{frag.layers.cohesion}

\title{Cohesion measure of layers}
\description{Returns the cohesion value of two archaeological spatial units (layers or any other type of spatial units).}
\usage{frag.layers.cohesion(graph, layer.attr, morphometry, x, y, z, verbose=TRUE)}

\arguments{
  \item{graph}{An undirected \code{igraph} object.} 
  \item{layer.attr}{Character. The name of the vertices attribute giving the spatial unit of the fragments.}
  \item{morphometry}{Character. Optional, to pass to the `frag.edges.weighting` function: name of the vertex attribute with the morphometric value of the fragments.}
  \item{x}{Character. Optional, to pass to the `frag.edges.weighting` function: name of the vertex attribute with the "x" coordinate of the fragments.}
  \item{y}{Character. Optional, to pass to the `frag.edges.weighting` function: name of the vertex attribute with the "y" coordinate of the fragments.}
  \item{z}{Character. Optional, to pass to the `frag.edges.weighting` function:  name of the vertex attribute with the "z" coordinate of the fragments.}
  \item{verbose}{Logical. Whether to print or not warning messages.}
}
\details{
Note that this function must be used after weighting the edges with \code{frag.edges.weighting}.



This function can be applied to three situations regarding the available information about the fragments location:
 \itemize{
    \item \bold{1 spatial unit}: The graph is compared to a minimized version of itself (i.e. same number of fragments and minimal number of edges). The \code{frag.edges.weighting} is applied internally.
    \item \bold{2 spatial units}: This is the basic intended use of this function, the graph edges must have been previously weighted with the \code{frag.edges.weighting} function (an error message is displayed if the graph does not have an edge attribute 'weight').
    \item \bold{>2 spatial units}: A subgraph is generated by pair of spatial units (using the \code{frag.get.layers.pair} function), the \code{frag.edges.weighting} function is automatically applied to each subgraph (a warning message is displayed), and pairwise cohesion are computed.
}
The computation takes into account the information imbalance between from each spatial unit.
Results range in ]0;1[, with values towards 0 for two spatial units with only inter-spatial units connection relationships, and values towards 1 if there is no relationship between spatial units and a significant imbalance of information about the two spatial units.

When applied to a \bold{pair} of spatial units, the cohesion value of a spatial unit is computed as: 

	\deqn{\frac{V_{unit_i} + W_{unit_i}}{ \sum_{j=1}^{2} V_{unit_j} + W_{unit_j}} }
	
with V the number of vertices in the unit and W the sum of the edge weights within the unit.


When applied to a \bold{single} spatial unit, the cohesion computed as above (comparing the observed graph with its minimized version) but the result is modified as:

	\deqn{(C - 0.5)  * 2}
    
with C the cohesion value obtained about the observed graph. Results range in [0;1] with 0 for a graph that cannot be minimized (i.e. made of pairs of fragments only) and 1 for a full graph. 


}
\value{

For graphs with one spatial unit, the function returns a numeric vector with a cohesion value in [0;1].
For graphs with more than one spatial unit, the function returns a matrix: rows correspond to the comparison of two spatial units (row names include the labels of the spatial units, alphanumerically ordered and separated by '/', e.g. 'unit1/unit2'), and columns give the cohesion values in ]0;1[ of the two spatial units (with column names 'cohesion1' for unit1 and 'cochesion2' for unit2).
}
\author{Sebastien Plutniak <sebastien.plutniak at posteo.net>}

\seealso{
\code{\link[archeofrag]{frag.edges.weighting}},
\code{\link[archeofrag]{frag.get.layers.pair}}
}
\examples{
g <- frag.simul.process(n.components=20, vertices=50, disturbance=.1)
frag.layers.cohesion(g, layer.attr="layer")
}
 
\keyword{measurement}
