\name{frag.observer.failure}
\alias{frag.observer.failure}

\title{Simulate the failure of an observer in determining the relationships between fragments.}
\description{Simulate the failure of an observer in determining the relationships between fragments.}
\usage{frag.observer.failure(graph, likelihoods)}

\arguments{
  \item{graph}{An undirected \code{igraph} object.} 
  \item{likelihoods}{Numerical vector of values in [0,1] giving the likelihood of not observing a relation between two fragments.}
}
\details{
In determining connection relationships between archaeological fragments, archaeologists often consider the likelihood they fail in identifying some of these relationships.
Given an initial fragmentation graph, this function aims to simulate the effects of such different likelihood values.

For each value in the \code{likelihoods} parameter, a new graph is generated by randomly removing from the input graph the given proportion of edges.
To generate a series of comparable graphs with different likelihood, the function internally reset the seed for random number generation, so ensuring that the edges will be removed in the same order when the function is run for multiple likelihood values.

An error message is displayed if at least one of the \code{likelihoods} values is < 0 or > 1.
}
\value{
A list of graphs.
}
\author{Sebastien Plutniak <sebastien.plutniak at posteo.net>}

\seealso{
\code{\link[base]{set.seed}}
}
\examples{
g <- frag.simul.process(n.components=20, vertices=50, disturbance=.15)
frag.observer.failure(graph=g, likelihoods=c(0.05, 0.1))
}
 
\keyword{graphs}
