\name{archeofrag-package}
\alias{archeofrag-package}
\alias{archeofrag}
\docType{package}
\title{Archeofrag: Refitting and Spatial Analysis in Archaeology}
\description{Methods to analyse fragmented objects in archeology using refitting relationships between fragments scattered in archeological spatial units (e.g. stratigraphic layers). Graphs and graph theory are used to model archeological observations. The package is mainly based on the 'igraph' package for graph analysis. Functions can: 1) create, manipulate, and simulate fragmentation graphs, 2) measure the cohesion and admixture of archeological spatial units, and 3) characterise the topology of a specific set of refitting relationships. An empirical dataset is also provided as an example.
}
\author{Sebastien Plutniak
Maintainer: Sebastien Plutniak <sebastien.plutniak at posteo.net>
}
\keyword{package}
\seealso{
\code{\link[igraph:igraph-package]{igraph}},
\code{\link[RBGL:RBGL.overview]{RBGL}} 
}
